refactor(commands): extract notes domain into commands/notes.rs (gated, verbatim)
Pure MOVE, zero behavior change. The 17 note-domain Tauri commands + their
note-only private helpers/consts lift out of commands/mod.rs into a new
commands/notes.rs (#[path]-aliased notes_commands, matching every precedent
domain); mod.rs 34,582 -> 33,652 (-930). lib.rs generate_handler! BYTE-FOR-BYTE
UNCHANGED.

ZERO pub(crate) promotions needed: a commands/*.rs submodule is a descendant
of commands, so it sees mod.rs's private fns via use super::* (same as the
already-merged analytics/meetings extractions), and every moved *_inner core a
retained test/caller reaches was already pub(crate) on trunk. Widened nothing.

Gate/mask/seal/export-gate logic UNCHANGED, only relocated (each verified a
verbatim substring of the new file):
- get_note -> masked_note_doc for sealed-not-unlocked (title locked, no body).
- list_notes / list_notes_typed -> gated Db::list_notes_visible* (sealed absent).
- create_note / update_note_doc / save_note_text / set_note_folder_schema:
  write-gate folder_is_unlocked -> AppError::Locked before touching plaintext;
  seal-on-write via sealed_document_blob (encrypt+VERIFY, fail-closed).
- move_note_doc_inner: seal-then-reassign + delete-then-move verify-before-destroy
  ordering.
- export_note_doc: gate FIRST + assert_in_vault; delete_note_inner: revoke-before-delete.
The shared DTO/mask helpers (note_display_title, note_doc_from_row, masked_note_doc)
and the vault-write leg (export_note_to_vault / write_note_to_vault) STAY in mod.rs
(used by share/org/unseal paths); notes.rs reaches them via super::.

Tests all stayed in mod.rs: #[test] 346 (= trunk); commands total 157 (= trunk).
Verified: cargo test --lib note 173/0, commands:: 351/0 (incl the masked-audio_path
asset-protocol gate + assert_in_vault + lifecycle relock-never-destroys), clippy
--lib -D warnings clean, MSRV grep clean.
