+++ src-tauri/src/commands/mod.rs
@@ -117,6 +117,15 @@ pub use ask_commands::*;
 mod meetings_commands;
 pub use meetings_commands::*;
 
+// Standalone-NOTES command surface (authored-note CRUD + Notes folder tree) — extracted verbatim
+// (God-file split, PURE MOVE — gate/mask/seal/export-gate bodies byte-identical, only relocated).
+// Bound as `notes_commands` (via `#[path]`) to keep it clearly distinct and avoid any future name
+// shadow with a `crate::notes`-style module. The glob re-export makes every moved command resolve
+// UNCHANGED at `crate::commands::…` for `generate_handler!` in `lib.rs` and every caller.
+#[path = "notes.rs"]
+mod notes_commands;
+pub use notes_commands::*;
+
 /// Keychain account for the AI Gateway API key (matches `summarize::GATEWAY_KEY_ACCOUNT`).
 /// Strictly separate from `ANTHROPIC_KEY_ACCOUNT` (defined in `commands/secrets.rs`) — never a
 /// fallback to the Anthropic key (R3). Kept here because the gateway model-listing helpers below
@@ -2935,606 +2944,6 @@ fn masked_note_doc(row: &crate::storage::db::NoteRow) -> NoteDoc {
     }
 }
 
-/// Create an EMPTY authored note. `folder_id = None` ⇒ the default note-folder (created on first
-/// use). WRITE-GATED: a sealed-and-not-session-unlocked target folder is refused (never resurrect
-/// plaintext behind a lock). Returns the new note id. (Unlike `import_text` the empty-text refusal
-/// is relaxed — an empty note is the whole point of "New note".)
-#[tauri::command]
-pub fn create_note(
-    state: State<'_, AppState>,
-    folder_id: Option<String>,
-    title: String,
-) -> Result<String, AppError> {
-    create_note_inner(state.inner(), folder_id.as_deref(), &title)
-}
-
-/// Inner of [`create_note`] taking `&AppState` (unit-testable gate).
-pub(crate) fn create_note_inner(
-    state: &AppState,
-    folder_id: Option<&str>,
-    title: &str,
-) -> Result<String, AppError> {
-    // BLK-1 / TOCTOU (2026-07-10 audit F4): hold the lifecycle guard across gate+insert so a
-    // concurrent lock/relock cannot land between the unlock check and the row insert.
-    let _lifecycle = lifecycle_guard(state);
-    // Resolve the anchor note-folder. A None/empty selection ⇒ the reserved always-open Notes ROOT
-    // (unfiled), NOT the old default "Notes" folder — so "New note" can NEVER fail because the default
-    // folder is sealed (the 2026-07-14 "Couldn't create the note" dead-end). `ensure_notes_root` is
-    // idempotent + never locked.
-    let folder_id = match folder_id {
-        Some(f) if !f.is_empty() => {
-            // Must be a NOTE folder — never create a note inside a meeting folder.
-            if state.db.note_folder_by_id(f)?.is_none() {
-                return Err(AppError::InvalidArg(format!("no note folder {f}")));
-            }
-            f.to_string()
-        }
-        _ => state.db.ensure_notes_root()?,
-    };
-
-    // WRITE-GATE: a sealed-and-not-session-unlocked folder is refused.
-    if !folder_is_unlocked(state, &folder_id)? {
-        return Err(AppError::Locked(
-            "this folder is locked — unlock it to add a note".into(),
-        ));
-    }
-
-    let title = title.trim();
-    let title = if title.is_empty() { "Untitled" } else { title };
-    // The `name` is the filesystem-safe slug; `title` is the display title.
-    let name = crate::export::sanitize_title(title);
-
-    let id = uuid::Uuid::new_v4().to_string();
-    let now = chrono::Utc::now().timestamp_millis();
-    // Seal-on-write (2026-07-10 audit F1, tightened by residual W3 to SEAL-THEN-INSERT): a note
-    // created in a session-unlocked LOCKED folder is sealed FROM BIRTH — the blob for the empty
-    // body is computed + VERIFIED first (fail-closed on a missing session KEK, BEFORE any row
-    // exists), then text + blob land in ONE atomic INSERT. The pre-fix insert-then-seal left a
-    // blob-less plaintext row lingering in the locked folder when the birth-seal failed (the
-    // relock reblank rightly refuses to blank a blob-less row — never destroy the only copy).
-    // Open folders: the plain insert.
-    let locked = state
-        .db
-        .folder_by_id(&folder_id)?
-        .map(|f| f.locked)
-        .unwrap_or(false);
-    if locked {
-        let blob = sealed_document_blob(state, &folder_id, &id, "")?;
-        state
-            .db
-            .insert_note_sealed(&id, &folder_id, &name, title, "", &blob, now)?;
-    } else {
-        state.db.insert_note(&id, &folder_id, &name, title, "", now)?;
-    }
-    // No chunks to index for an empty body — `update_note` re-indexes once the user writes.
-    // Brain v3 PR-3 — LINK ENGINE: index the (empty) body's wikilinks so a create with no body
-    // establishes a clean empty edge set; the first `update_note_doc` re-indexes real `[[Title]]`s.
-    // No semantic pass on birth (no vectors yet). Best-effort.
-    index_wikilinks_best_effort(state, crate::links::LinkKind::Note, &id, "");
-    tracing::info!(target: "notes", note_id = %id, folder_id = %folder_id, "note created");
-    Ok(id)
-}
-
-/// System prompt for the on-device auto-title (Feature B, 2026-07-14). Short + strict output so the
-/// result is a clean headline, not a sentence.
-const AUTO_TITLE_SYSTEM: &str = "You write a short, specific title (3 to 6 words) for a note. \
-Output ONLY the title text — no surrounding quotes, no trailing punctuation, no 'Title:' prefix.";
-
-/// The on-device summarizer for a LOCAL-ONLY, zero-egress title — `None` when the on-device model
-/// isn't downloaded (the caller then uses the first-line heuristic). NEVER the cloud: an auto-title
-/// must never egress the note. Mirrors the local-provider construction in `summarize::provider_for`.
-fn local_title_provider(
-    state: &AppState,
-) -> Option<std::sync::Arc<dyn crate::summarize::provider::SummarizerProvider>> {
-    let (configured_path, model_id, timeouts) = {
-        let config = state.config.lock().ok()?;
-        (
-            config.brain_model_path.clone(),
-            config.brain_model_id.clone(),
-            crate::reason::sidecar::SidecarTimeouts {
-                idle_secs: config.brain_idle_timeout_secs,
-                ready_secs: config.brain_ready_timeout_secs,
-                hard_cap_secs: config.brain_hard_cap_secs,
-            },
-        )
-    };
-    let configured = configured_path.as_deref().map(std::path::Path::new);
-    let path = crate::reason::resolve_brain_model(configured, model_id.as_deref()).ok()??;
-    let reasoner: std::sync::Arc<dyn crate::reason::LocalReasoner> =
-        std::sync::Arc::new(crate::reason::sidecar::SidecarReasoner::new(path, timeouts).ok()?);
-    Some(std::sync::Arc::new(
-        crate::summarize::local::LocalSummarizerProvider::new(
-            reasoner,
-            std::sync::Arc::clone(&state.heavy_inference),
-        ),
-    ))
-}
-
-/// Clean an LLM title suggestion into a single tidy line, or `None` if it collapsed to nothing.
-fn sanitize_title_suggestion(raw: &str) -> Option<String> {
-    let line = raw.trim().lines().next().unwrap_or("").trim();
-    let line = line
-        .trim_matches('"')
-        .trim_matches('\'')
-        .trim_start_matches('#')
-        .trim();
-    let line = line.trim_end_matches(['.', ':', '-']).trim();
-    if line.is_empty() {
-        return None;
-    }
-    Some(line.chars().take(80).collect())
-}
-
-/// Fallback title: the first non-empty line of the body, stripped of leading markdown, capped.
-fn first_line_title(body: &str) -> String {
-    let line = body
-        .lines()
-        .map(|l| l.trim())
-        .find(|l| !l.is_empty())
-        .unwrap_or("");
-    let line = line.trim_start_matches(['#', '-', '*', '>', ' ']).trim();
-    let capped: String = line.chars().take(60).collect();
-    if capped.is_empty() {
-        "Untitled".to_string()
-    } else {
-        capped
-    }
-}
-
-/// `suggest_note_title(note_id)` — auto-title an "Untitled" note from its body (Feature B): the
-/// on-device model when present, else a first-line heuristic. LOCAL-ONLY (never egresses). Persists
-/// the new title ONLY if the note is still "Untitled" (never clobbers a user's title) and skips a note
-/// in a LOCKED folder (spec decision #3 — no auto-unlock / seal-on-write for a title). Returns the
-/// title (the current one unchanged when it skips). Called fire-and-forget by the editor on close.
-#[tauri::command]
-pub async fn suggest_note_title(
-    state: State<'_, AppState>,
-    note_id: String,
-) -> Result<String, AppError> {
-    suggest_note_title_inner(state.inner(), &note_id).await
-}
-
-pub(crate) async fn suggest_note_title_inner(
-    state: &AppState,
-    note_id: &str,
-) -> Result<String, AppError> {
-    let Some(row) = state.db.get_note_row(note_id)? else {
-        return Err(AppError::InvalidArg(format!("no note {note_id}")));
-    };
-    // `NoteRow.title` is Option<String> (NULL-able column) — treat a missing title as "".
-    let current = row.title.clone().unwrap_or_default();
-    // Only ever fill in an "Untitled" note — never overwrite a title the user chose.
-    let cur = current.trim();
-    if !cur.is_empty() && cur != "Untitled" {
-        return Ok(current);
-    }
-    // Skip a locked folder: new notes live in the always-open root, so this only skips the rare
-    // deliberately-filed-then-locked case; auto-titling it would need seal-on-write / an unlock.
-    let folder_locked = state
-        .db
-        .folder_by_id(&row.folder_id)?
-        .map(|f| f.locked)
-        .unwrap_or(false);
-    if folder_locked {
-        return Ok(current);
-    }
-    let body = row.text.trim();
-    if body.is_empty() {
-        return Ok(current);
-    }
-
-    // On-device first (zero egress); first-line heuristic otherwise / on failure.
-    let title = match local_title_provider(state) {
-        Some(provider) => {
-            let excerpt: String = body.chars().take(1200).collect();
-            match provider.complete(AUTO_TITLE_SYSTEM, &excerpt).await {
-                Ok(t) => sanitize_title_suggestion(&t).unwrap_or_else(|| first_line_title(body)),
-                Err(e) => {
-                    tracing::info!(target: "notes", error = %e, "auto-title: on-device failed, using first line");
-                    first_line_title(body)
-                }
-            }
-        }
-        None => first_line_title(body),
-    };
-    if title.trim().is_empty() || title == "Untitled" {
-        return Ok(current);
-    }
-
-    // Persist — re-read + re-check under the write so a title the user set meanwhile always wins.
-    if let Some(fresh) = state.db.get_note_row(note_id)? {
-        let fresh_title = fresh.title.clone().unwrap_or_default();
-        let ft = fresh_title.trim();
-        if ft.is_empty() || ft == "Untitled" {
-            let now = chrono::Utc::now().timestamp_millis();
-            state.db.update_note_row(note_id, &title, &fresh.text, now)?;
-            return Ok(title);
-        }
-        return Ok(fresh_title); // user titled it in the meantime
-    }
-    Ok(title)
-}
-
-/// Read ONE note (editor DTO). GATED: a sealed-and-not-session-unlocked note returns the MASKED DTO
-/// (title "🔒 Locked", no body/tags), never the stored text.
-#[tauri::command]
-pub fn get_note(state: State<'_, AppState>, id: String) -> Result<NoteDoc, AppError> {
-    get_note_inner(state.inner(), &id)
-}
-
-/// Inner of [`get_note`] taking `&AppState` (unit-testable gate).
-pub(crate) fn get_note_inner(state: &AppState, id: &str) -> Result<NoteDoc, AppError> {
-    let Some(row) = state.db.get_note_row(id)? else {
-        return Err(AppError::InvalidArg(format!("no note {id}")));
-    };
-    if !folder_is_unlocked(state, &row.folder_id)? {
-        return Ok(masked_note_doc(&row)); // sealed-not-unlocked ⇒ masked, never the stored text.
-    }
-    let mut doc = note_doc_from_row(&row);
-    doc.shared = state.db.note_has_active_share(&row.id)?; // WP6 — active-share flag.
-    Ok(doc)
-}
-
-/// List note summaries (leak-free). `folder_id = None` ⇒ all VISIBLE notes; `Some(fid)` scopes to
-/// one note-folder. Filtered IN THE QUERY by `visibility_clause` — a sealed-not-unlocked note is
-/// ABSENT (its title never leaks), not per-row masked.
-#[tauri::command]
-pub fn list_notes(
-    state: State<'_, AppState>,
-    folder_id: Option<String>,
-) -> Result<Vec<NoteSummary>, AppError> {
-    list_notes_inner(state.inner(), folder_id.as_deref())
-}
-
-/// Inner of [`list_notes`] taking `&AppState` (unit-testable gate).
-pub(crate) fn list_notes_inner(
-    state: &AppState,
-    folder_id: Option<&str>,
-) -> Result<Vec<NoteSummary>, AppError> {
-    let unlocked = unlocked_snapshot(state)?;
-    state.db.list_notes_visible(folder_id, &unlocked)
-}
-
-/// Update a note's title + markdown (write path). WRITE-GATED. Bumps `updated_at`, PURGES the old
-/// `doc_chunks` and re-chunks+re-embeds the new BODY (front-matter stripped) so the note stays a
-/// first-class brain source, then re-exports the vault `.md`. Returns the fresh DTO.
-///
-/// COMMIT BOUNDARY (authored-note full save): after the local write succeeds, BEST-EFFORT re-publish
-/// any org shares of this note (never freeze the org copy at share time). NOT the debounced
-/// `save_note_text` autosave (OCK-seal + egress per keystroke is unacceptable) — only this commit path.
-#[tauri::command]
-pub async fn update_note_doc(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    id: String,
-    title: String,
-    markdown: String,
-) -> Result<NoteDoc, AppError> {
-    // PERF (brain-v3 audit H3): the inner runs Candle/Metal embedding + the semantic-link pass —
-    // route the whole synchronous body through the shared heavy-inference gate on the blocking
-    // pool, exactly like the meeting twin `update_note` (PR-1 #362). Re-fetch `AppState` inside
-    // the closure via `app.state()` — a bare `&AppState` cannot be captured by a `'static` closure.
-    let heavy_inference = state.heavy_inference.clone();
-    let app_for_edit = app.clone();
-    let id_for_edit = id.clone();
-    let title_for_edit = title.clone();
-    let markdown_for_edit = markdown.clone();
-    let doc = crate::perf::run_heavy(&heavy_inference, move || -> Result<NoteDoc, AppError> {
-        let state = app_for_edit.state::<AppState>();
-        update_note_doc_inner(&state, &id_for_edit, &title_for_edit, &markdown_for_edit)
-    })
-    .await?;
-    // See `update_note`: ping open org views when the edit re-published ≥1 org copy. Best-effort.
-    if republish_org_shares_for_source(state.inner(), None, Some(&id))
-        .await
-        .unwrap_or(0)
-        > 0
-    {
-        crate::events::emit_org_feed_updated(&app, 1);
-    }
-    Ok(doc)
-}
-
-/// Inner of [`update_note_doc`] taking `&AppState` (unit-testable gate). Resolves the model-gated
-/// embedder and delegates to [`update_note_doc_inner_with`] (embedder injected for deterministic
-/// tests — the [`update_note_inner`] precedent).
-pub(crate) fn update_note_doc_inner(
-    state: &AppState,
-    id: &str,
-    title: &str,
-    markdown: &str,
-) -> Result<NoteDoc, AppError> {
-    let embedder = crate::embed::embed_model_present().then(crate::embed::active_embedder);
-    update_note_doc_inner_with(state, id, title, markdown, embedder.as_deref())
-}
-
-/// Core of [`update_note_doc_inner`] with the re-index embedder INJECTED (`None` = model absent →
-/// chunk-only re-index, FTS still covers keyword retrieval; `Some` = fresh vectors too).
-pub(crate) fn update_note_doc_inner_with(
-    state: &AppState,
-    id: &str,
-    title: &str,
-    markdown: &str,
-    embedder: Option<&dyn crate::embed::Embedder>,
-) -> Result<NoteDoc, AppError> {
-    // BLK-1 / TOCTOU (2026-07-10 audit F4): hold the lifecycle guard across gate+write so a
-    // concurrent relock/seal cannot land between the unlock check and the row write.
-    let lifecycle = lifecycle_guard(state);
-    let Some(row) = state.db.get_note_row(id)? else {
-        return Err(AppError::InvalidArg(format!("no note {id}")));
-    };
-    // WRITE-GATE: refuse editing a sealed-and-not-session-unlocked note (never write plaintext
-    // behind a lock).
-    if !folder_is_unlocked(state, &row.folder_id)? {
-        return Err(AppError::Locked(
-            "this note is locked — unlock its folder to edit it".into(),
-        ));
-    }
-    let title = title.trim();
-    let title = if title.is_empty() { "Untitled" } else { title };
-    let now = chrono::Utc::now().timestamp_millis();
-    // Seal-on-write (audit F1): a session-unlocked LOCKED folder re-seals the fresh text into
-    // `text_blob` in the same write; an open folder takes the plain update. Fail-closed on a
-    // missing session KEK.
-    reseal_document_if_locked(state, &row.folder_id, id, title, markdown, now)?;
-
-    // Re-export the vault `.md` (best-effort). A sealed folder has no export (gated above), so this
-    // only runs for a visible note.
-    if let Err(e) = export_note_to_vault(state, id) {
-        tracing::warn!(target: "notes", error = %e, "note vault re-export failed (text saved)");
-    }
-    // Read the fresh DTO while still guarded (the row cannot be resealed under us); the heavy leg
-    // below never mutates the note row, so the DTO stays accurate.
-    let doc = get_note_inner(state, id)?;
-
-    // PERF (brain-v3 audit H3, the `update_note_inner_with` twin): release the GLOBAL lifecycle
-    // guard BEFORE the heavy leg — the note-body re-embed + semantic-link pass is multi-second
-    // Candle/Metal work on a long note, and holding the guard across it delays screen-share
-    // auto-relock and every lock/unlock op. Race-safe without it: `index_note_chunks` re-checks
-    // the sealed-at-rest invariant (`doc_sealed_at_rest_tx`) INSIDE its write tx, so a mid-flight
-    // seal makes the write a refused no-op. The fast leg above (gate + seal-on-write + vault
-    // export + DTO read) stays under the guard.
-    drop(lifecycle);
-
-    // WP3 — the note is a first-class brain source: purge the old chunks and re-index the new BODY
-    // (front-matter stripped inside `index_note_body_chunks`). Chunks + FTS come back
-    // unconditionally (keyword works model-less); vectors ONLY when the REAL e5 model is present
-    // (never stub vectors; mirrors `ingest_into_folder`/`should_auto_index`). Best-effort: a failure
-    // logs (no PII) and does NOT fail the update — the plaintext is durable.
-    if let Err(e) = index_note_body_chunks(state, id, title, markdown, embedder) {
-        tracing::warn!(target: "rag", error = %e, "note re-index on update failed (text saved)");
-    }
-
-    // Brain v3 PR-3 — LINK ENGINE: (a) re-index this note's `[[Title]]` wikilink edges by resolved
-    // TARGET id (rename-proof), and (b) refresh its semantic suggestions from the fresh vectors
-    // (model-gated). Both best-effort — a link failure never fails the note save.
-    index_wikilinks_best_effort(state, crate::links::LinkKind::Note, id, markdown);
-    auto_link_semantic_best_effort(state, crate::links::LinkKind::Note, id);
-
-    tracing::info!(target: "notes", note_id = %id, "note updated + re-indexed");
-    Ok(doc)
-}
-
-/// FAST autosave path — persist the note's title + markdown + `updated_at` ONLY. NO re-chunk, NO
-/// re-embed, NO vault re-export: those are the expensive parts (the e5 embedder saturates the
-/// machine if run on every keystroke-pause) and are DEFERRED to [`update_note_doc`], which the FE
-/// runs on the natural boundaries (editor blur/close, Preview, explicit export/share). The frequent
-/// debounced autosave calls THIS so typing stays smooth even with the embed model present. WRITE-
-/// GATED exactly like the full update — a sealed-and-not-session-unlocked note is refused (never
-/// write plaintext behind a lock). The DB text is canonical, so nothing is lost if the app closes
-/// before the next full update. Returns the new `updated_at` (epoch ms) so the FE reconciles its
-/// indicator without a full DTO round-trip.
-#[tauri::command]
-pub fn save_note_text(
-    state: State<'_, AppState>,
-    id: String,
-    title: String,
-    markdown: String,
-) -> Result<i64, AppError> {
-    save_note_text_inner(state.inner(), &id, &title, &markdown)
-}
-
-/// Inner of [`save_note_text`] taking `&AppState` (unit-testable gate).
-pub(crate) fn save_note_text_inner(
-    state: &AppState,
-    id: &str,
-    title: &str,
-    markdown: &str,
-) -> Result<i64, AppError> {
-    // BLK-1 / TOCTOU (2026-07-10 audit F4): hold the lifecycle guard across gate+write so a
-    // concurrent relock/seal cannot land between the unlock check and the row write.
-    let _lifecycle = lifecycle_guard(state);
-    let Some(row) = state.db.get_note_row(id)? else {
-        return Err(AppError::InvalidArg(format!("no note {id}")));
-    };
-    if !folder_is_unlocked(state, &row.folder_id)? {
-        return Err(AppError::Locked(
-            "this note is locked — unlock its folder to edit it".into(),
-        ));
-    }
-    let title = title.trim();
-    let title = if title.is_empty() { "Untitled" } else { title };
-    let now = chrono::Utc::now().timestamp_millis();
-    // Seal-on-write (audit F1): a session-unlocked LOCKED folder re-seals the fresh text into
-    // `text_blob` in the same write; an open folder takes the plain update. Fail-closed on a
-    // missing session KEK.
-    reseal_document_if_locked(state, &row.folder_id, id, title, markdown, now)?;
-    Ok(now)
-}
-
-/// Move a note to a different note-folder. GATED on BOTH the source and target folder-unlock (never
-/// leave plaintext in a locked folder, never move out of a sealed one). Re-exports into the new
-/// folder path (the old vault file is removed by the export path's move). Idempotent.
-#[tauri::command]
-pub fn move_note_doc(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    id: String,
-    folder_id: String,
-) -> Result<(), AppError> {
-    move_note_doc_inner(state.inner(), &id, &folder_id)?;
-    // A move INTO a locked folder seals + purges ALL pending audit findings; an open-target move
-    // purges nothing — the count-only ping is correct (and cheap) either way.
-    emit_audit_updated_after_purge(&app, state.inner());
-    Ok(())
-}
-
-/// Inner of [`move_note_doc`] taking `&AppState` (unit-testable gate).
-pub(crate) fn move_note_doc_inner(
-    state: &AppState,
-    id: &str,
-    folder_id: &str,
-) -> Result<(), AppError> {
-    // BLK-1 / TOCTOU (2026-07-10 audit F4): hold the lifecycle guard across the double gate + the
-    // reassign/seal writes so a concurrent lock/relock cannot land mid-move.
-    let _lifecycle = lifecycle_guard(state);
-    let Some(row) = state.db.get_note_row(id)? else {
-        return Err(AppError::InvalidArg(format!("no note {id}")));
-    };
-    // Source gate: refuse moving a sealed-not-unlocked note (its plaintext is blanked — we'd move an
-    // empty note and lose the seal anchoring; unlock first).
-    if !folder_is_unlocked(state, &row.folder_id)? {
-        return Err(AppError::Locked(
-            "this note is locked — unlock its folder to move it".into(),
-        ));
-    }
-    // Target must be a NOTE folder and be unlocked (never land plaintext behind a lock). We do not
-    // support sealing-on-move into a locked note-folder in WP0 — refuse (the FE unlocks first).
-    if state.db.note_folder_by_id(folder_id)?.is_none() {
-        return Err(AppError::InvalidArg(format!("no note folder {folder_id}")));
-    }
-    if !folder_is_unlocked(state, folder_id)? {
-        return Err(AppError::Locked(
-            "the target folder is locked — unlock it to move a note there".into(),
-        ));
-    }
-
-    // Seal alignment (2026-07-10 audit F1, reordered by residual W2 to SEAL-THEN-REASSIGN): the
-    // seal blob must always match the OWNING folder's CK. Moving INTO a session-unlocked LOCKED
-    // folder seals the plaintext under the TARGET folder's CK FIRST — encrypt + verify (fail-closed
-    // on a missing session KEK) BEFORE anything is mutated — then reassigns folder + fresh blob +
-    // NULLed exported_path in ONE atomic UPDATE ([`Db::move_note_row_sealed`]). The pre-fix
-    // reassign-then-seal left the note sitting in the locked target with a stale wrong-CK blob when
-    // the reseal failed (the target's next unlock would fail on it); now a seal failure leaves the
-    // note untouched in the SOURCE folder (old `.md` included — the seal is computed BEFORE the
-    // R1 delete below, so a fail-closed seal mutates nothing at all).
-    // Moving into an OPEN folder clears any stale blob from the previous folder's CK (the plaintext
-    // column is canonical again — a stale blob would be undecryptable under a future lock of the
-    // new folder and could shadow fresh content).
-    //
-    // R1 (#231 review residual) — DELETE-THEN-MOVE: the OLD vault `.md` is removed BEFORE the
-    // move UPDATE clears/NULLs `exported_path` (but AFTER the fallible seal, so a seal failure
-    // still touches nothing). INVARIANT: a plaintext `.md` of locked-folder content is never left
-    // on disk with no `exported_path` record pointing at it — the pre-fix best-effort remove
-    // AFTER the UPDATE orphaned the file when the delete failed (or the app crashed between the
-    // two), untracked and unreconcilable. An already-absent file is success; any other delete
-    // failure REFUSES the move with the row (folder, exported_path, seal state) untouched — the
-    // user retries. The DB `text` column stays the canonical copy throughout, so a crash after
-    // this delete loses no content (the re-export recreates the `.md`).
-    let target_locked = state
-        .db
-        .folder_by_id(folder_id)?
-        .map(|f| f.locked)
-        .unwrap_or(false);
-    if target_locked {
-        let title = row.title.clone().unwrap_or_else(|| row.name.clone());
-        let updated_at = row.updated_at.unwrap_or(row.created_at);
-        let blob = sealed_document_blob(state, folder_id, id, &row.text)?;
-        remove_note_export_before_move(row.exported_path.as_deref())?;
-        state
-            .db
-            .move_note_row_sealed(id, folder_id, &title, &row.text, &blob, updated_at)?;
-        // Vault Audit LOCK-SAFETY: a move-into-locked is a SEAL — purge ALL pending findings
-        // (the rollup posture; this note's title may be cited in third-party evidence no id can
-        // match). The other seal paths purge inside their chunk-purge txs; this path has none.
-        state.db.purge_all_pending_audit_findings()?;
-    } else {
-        remove_note_export_before_move(row.exported_path.as_deref())?;
-        state.db.set_note_doc_folder(id, folder_id)?;
-        state.db.set_note_doc_exported_path(id, None)?;
-        if row.sealed {
-            state.db.clear_document_blob(id)?;
-        }
-    }
-    if let Err(e) = export_note_to_vault(state, id) {
-        tracing::warn!(target: "notes", error = %e, "note re-export after move failed (moved in db)");
-    }
-    tracing::info!(target: "notes", note_id = %id, folder_id = %folder_id, "note moved");
-    Ok(())
-}
-
-/// R1 helper — delete a note's OLD exported vault `.md` as the DELETE-THEN-MOVE step of
-/// [`move_note_doc_inner`]. `None` / an already-absent file is success; any other failure is a
-/// clear [`AppError::Export`] the caller propagates to REFUSE the move (nothing mutated yet, the
-/// user retries) — never a silently orphaned plaintext file.
-fn remove_note_export_before_move(old_path: Option<&str>) -> Result<(), AppError> {
-    let Some(old_path) = old_path else {
-        return Ok(()); // never exported — nothing to remove.
-    };
-    match std::fs::remove_file(old_path) {
-        Ok(()) => Ok(()),
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(()),
-        Err(e) => Err(AppError::Export(format!(
-            "could not remove the note's exported vault file before moving it: {e}"
-        ))),
-    }
-}
-
-/// Permanently delete a note (cascade its chunks + vectors + vault `.md`). GATED: a
-/// sealed-and-not-session-unlocked folder is refused. Reuses `delete_document` semantics.
-///
-/// DELETE-CASCADE FIX (2026-07-15): the local hard-delete used to leave any live org share of this
-/// note `uploaded` — the server ciphertext survived + the 60s background org-sync tick re-pulled it
-/// back into the local replica, resurrecting a "deleted" note. Now revokes every live org share of
-/// this exact note FIRST (fails loud on a revoke error — see `revoke_org_shares_for_source`), then
-/// fans out a content-free delete event so any other open surface (the tab-strip) can prune itself.
-/// `async` (was sync) because the revoke is a network round-trip.
-#[tauri::command]
-pub async fn delete_note(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    id: String,
-) -> Result<(), AppError> {
-    delete_note_inner(state.inner(), &id).await?;
-    crate::events::emit_content_deleted(&app, "note", &id);
-    // The delete purged its audit findings (id-matched) — ping the FE inbox (count-only).
-    emit_audit_updated_after_purge(&app, state.inner());
-    Ok(())
-}
-
-/// Inner of [`delete_note`] taking `&AppState` (unit-testable gate). `async` for the org-share revoke
-/// cascade (network round-trip); the gate + DB delete themselves stay synchronous internally.
-pub(crate) async fn delete_note_inner(state: &AppState, id: &str) -> Result<(), AppError> {
-    let Some(row) = state.db.get_note_row(id)? else {
-        return Ok(()); // unknown id → idempotent no-op.
-    };
-    if !folder_is_unlocked(state, &row.folder_id)? {
-        return Err(AppError::Locked(
-            "this folder is locked — unlock it to delete a note".into(),
-        ));
-    }
-    // REVOKE-BEFORE-DELETE (Bug A root cause): tear down every LIVE org share of this exact note
-    // BEFORE the local row disappears, so the background org-sync tick can never re-pull a still-live
-    // server item back into the local replica after the user asked to delete it. Fails LOUD: a revoke
-    // failure (e.g. offline) aborts the delete rather than silently leaving a dangling live share.
-    revoke_org_shares_for_source(state, None, Some(id)).await?;
-    // Remove the vault file first (best-effort), then cascade-delete the row + its chunks/vectors.
-    if let Some(path) = &row.exported_path {
-        let _ = std::fs::remove_file(path);
-    }
-    state.db.delete_document(id)?;
-    tracing::info!(target: "notes", note_id = %id, "note deleted");
-    Ok(())
-}
-
-/// (Re)write a note's vault `.md` and return the path. GATED. Idempotent (atomic, collision-suffixed
-/// `write_note`). Stores the path in `documents.exported_path`.
-#[tauri::command]
-pub fn export_note_doc(state: State<'_, AppState>, id: String) -> Result<String, AppError> {
-    let path = export_note_to_vault(state.inner(), &id)?;
-    path.ok_or_else(|| AppError::Export("no vault configured — set your Obsidian vault first".into()))
-}
-
 /// WP1 — write the note's markdown to `<vault>/<note-folder-path>/<title>.md` (note folders are
 /// rooted under `Notes/…`, so this lands under `<vault>/Notes/…`), record the path in
 /// `exported_path`. GATED (a sealed-not-unlocked note is never exported). Returns `Ok(None)` when no
@@ -3624,345 +3033,6 @@ fn index_note_body_chunks(
     state
         .db
         .index_note_chunks(id, title, &body, embedder)
-}
-
-/// List every note-folder (`kind='note'`). Lock state comes through unchanged from `folders.locked`.
-#[tauri::command]
-pub fn list_note_folders(state: State<'_, AppState>) -> Result<Vec<NoteFolder>, AppError> {
-    let mut folders = state.db.list_note_folders()?;
-    // Join the live session unlock set so a sealed-but-session-unlocked note folder reports
-    // `unlocked: true` — mirrors `list_folders` for the Meetings tree. Without this the Notes
-    // lock gate reads only the DB `locked` column (which never flips on session-unlock), so
-    // "Unlock folder" appeared to do nothing (the gate never lifted). (2026-07-14 fix.)
-    let unlocked = {
-        state
-            .unlocked_folders
-            .lock()
-            .map_err(|_| AppError::Storage("unlocked-folders mutex poisoned".into()))?
-            .clone()
-    };
-    for f in &mut folders {
-        f.unlocked = f.locked && unlocked.contains(&f.id);
-    }
-    Ok(folders)
-}
-
-// ── Feature C — TYPED note front-matter properties (note-folder schemas + Table/Board substrate) ──
-
-/// Max property columns a note-folder schema may declare (a UI-scale bound, not a hard DB limit).
-const NOTE_SCHEMA_MAX_FIELDS: usize = 40;
-/// Max length of a schema property key.
-const NOTE_SCHEMA_MAX_KEY_LEN: usize = 60;
-
-/// Read a note-folder's typed property SCHEMA (Feature C). GATED (the SAFER choice): a
-/// sealed-and-not-session-unlocked folder returns `Ok(vec![])` — we deliberately do NOT expose a
-/// locked folder's schema (the schema is only needed to view/edit an UNLOCKED folder). `InvalidArg`
-/// when `folder_id` is not a note-folder.
-#[tauri::command]
-pub fn get_note_folder_schema(
-    state: State<'_, AppState>,
-    folder_id: String,
-) -> Result<Vec<PropertySchemaField>, AppError> {
-    get_note_folder_schema_inner(state.inner(), &folder_id)
-}
-
-/// Inner of [`get_note_folder_schema`] taking `&AppState` (unit-testable gate).
-pub(crate) fn get_note_folder_schema_inner(
-    state: &AppState,
-    folder_id: &str,
-) -> Result<Vec<PropertySchemaField>, AppError> {
-    if state.db.note_folder_by_id(folder_id)?.is_none() {
-        return Err(AppError::InvalidArg(format!("no note folder {folder_id}")));
-    }
-    // GATE: a locked-and-not-session-unlocked folder's schema is NOT exposed (return empty).
-    if !folder_is_unlocked(state, folder_id)? {
-        return Ok(Vec::new());
-    }
-    state.db.get_note_folder_schema(folder_id)
-}
-
-/// Set a note-folder's typed property SCHEMA (Feature C). WRITE-GATED: a sealed-and-not-session-
-/// unlocked folder is refused (`AppError::Locked`) — never mutate metadata behind a lock. Validates:
-/// ≤40 fields, each `key` non-empty/trimmed/≤60 chars, case-insensitively UNIQUE, never the reserved
-/// `"tags"`, and every `Select` field carries ≥1 non-empty option.
-#[tauri::command]
-pub fn set_note_folder_schema(
-    state: State<'_, AppState>,
-    folder_id: String,
-    fields: Vec<PropertySchemaField>,
-) -> Result<(), AppError> {
-    set_note_folder_schema_inner(state.inner(), &folder_id, fields)
-}
-
-/// Inner of [`set_note_folder_schema`] taking `&AppState` (unit-testable gate + validation).
-pub(crate) fn set_note_folder_schema_inner(
-    state: &AppState,
-    folder_id: &str,
-    fields: Vec<PropertySchemaField>,
-) -> Result<(), AppError> {
-    if state.db.note_folder_by_id(folder_id)?.is_none() {
-        return Err(AppError::InvalidArg(format!("no note folder {folder_id}")));
-    }
-    // WRITE-GATE (before any validation or write): a sealed-not-unlocked folder is refused so a
-    // schema can never be edited behind a lock.
-    if !folder_is_unlocked(state, folder_id)? {
-        return Err(AppError::Locked(
-            "this folder is locked — unlock it to edit its properties".into(),
-        ));
-    }
-    if fields.len() > NOTE_SCHEMA_MAX_FIELDS {
-        return Err(AppError::InvalidArg(format!(
-            "too many properties (max {NOTE_SCHEMA_MAX_FIELDS})"
-        )));
-    }
-    let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-    for f in &fields {
-        let key = f.key.trim();
-        if key.is_empty() {
-            return Err(AppError::InvalidArg("property key must not be empty".into()));
-        }
-        if key.len() > NOTE_SCHEMA_MAX_KEY_LEN {
-            return Err(AppError::InvalidArg(format!(
-                "property key too long (max {NOTE_SCHEMA_MAX_KEY_LEN} chars)"
-            )));
-        }
-        // `tags` is reserved for the front-matter tag list (parsed separately, never a property).
-        if key.eq_ignore_ascii_case("tags") {
-            return Err(AppError::InvalidArg(
-                "property key \"tags\" is reserved".into(),
-            ));
-        }
-        if !seen.insert(key.to_ascii_lowercase()) {
-            return Err(AppError::InvalidArg(format!(
-                "duplicate property key \"{key}\""
-            )));
-        }
-        if f.kind == PropertyKind::Select
-            && !f.options.iter().any(|o| !o.trim().is_empty())
-        {
-            return Err(AppError::InvalidArg(format!(
-                "select property \"{key}\" needs at least one option"
-            )));
-        }
-    }
-    // Persist the TRIMMED keys (so the read-time coercion keys match the front-matter keys exactly).
-    let normalized: Vec<PropertySchemaField> = fields
-        .into_iter()
-        .map(|f| PropertySchemaField {
-            key: f.key.trim().to_string(),
-            kind: f.kind,
-            options: f
-                .options
-                .into_iter()
-                .map(|o| o.trim().to_string())
-                .filter(|o| !o.is_empty())
-                .collect(),
-        })
-        .collect();
-    state.db.set_note_folder_schema(folder_id, &normalized)?;
-    tracing::info!(target: "notes", folder_id = %folder_id, fields = normalized.len(), "note-folder schema updated");
-    Ok(())
-}
-
-/// List a note-folder's notes projected through its typed schema (Feature C — the Table/Board
-/// substrate). GATED: a sealed-and-not-session-unlocked folder returns `[]` (no rows, never a masked
-/// row) via [`Db::list_notes_visible_typed`], which is built on the gated `list_notes_visible`.
-/// `InvalidArg` when `folder_id` is not a note-folder.
-#[tauri::command]
-pub fn list_notes_typed(
-    state: State<'_, AppState>,
-    folder_id: String,
-) -> Result<Vec<TypedNoteRow>, AppError> {
-    list_notes_typed_inner(state.inner(), &folder_id)
-}
-
-/// Inner of [`list_notes_typed`] taking `&AppState` (unit-testable gate).
-pub(crate) fn list_notes_typed_inner(
-    state: &AppState,
-    folder_id: &str,
-) -> Result<Vec<TypedNoteRow>, AppError> {
-    if state.db.note_folder_by_id(folder_id)?.is_none() {
-        return Err(AppError::InvalidArg(format!("no note folder {folder_id}")));
-    }
-    let unlocked = unlocked_snapshot(state)?;
-    state.db.list_notes_visible_typed(folder_id, &unlocked)
-}
-
-/// Create a note-folder (`kind='note'`). `parent_id` must itself be a note-folder (or None = a
-/// root note-folder). Path is rooted under the default "Notes" tree so it can't collide with a
-/// meeting-folder path. Mirrors [`create_folder`] but stamps kind='note'.
-#[tauri::command]
-pub fn create_note_folder(
-    state: State<'_, AppState>,
-    name: String,
-    parent_id: Option<String>,
-) -> Result<NoteFolder, AppError> {
-    create_note_folder_inner(state.inner(), &name, parent_id.as_deref())
-}
-
-/// Inner of [`create_note_folder`] taking `&AppState`.
-pub(crate) fn create_note_folder_inner(
-    state: &AppState,
-    name: &str,
-    parent_id: Option<&str>,
-) -> Result<NoteFolder, AppError> {
-    let clean = crate::summarize::organize::sanitize_folder(name)
-        .ok_or_else(|| AppError::InvalidArg("folder name is empty or invalid".into()))?;
-
-    // Resolve the parent's path. A None parent roots under the default "Notes" folder so every note
-    // folder is namespaced under "Notes/…" (can't collide with a meeting path; vault export lands
-    // under <vault>/Notes/…). A Some parent MUST be a note-folder.
-    let parent_path = match parent_id {
-        Some(pid) => {
-            let parent = state
-                .db
-                .note_folder_by_id(pid)?
-                .ok_or_else(|| AppError::InvalidArg(format!("no parent note folder {pid}")))?;
-            parent.path
-        }
-        None => {
-            // Ensure the root "Notes" folder exists and nest under it.
-            state.db.ensure_default_note_folder()?;
-            "Notes".to_string()
-        }
-    };
-    let rel_path = format!("{parent_path}/{clean}");
-
-    // Create the vault subdirectory (only when a vault is configured); assert it stays in-vault.
-    if let Some(vault) = vault_path(state) {
-        let vault_root = std::path::Path::new(&vault);
-        let dir = assert_in_vault(vault_root, std::path::Path::new(&rel_path))?;
-        std::fs::create_dir_all(&dir)
-            .map_err(|e| AppError::Export(format!("create note folder dir failed: {e}")))?;
-    }
-
-    let folder = NoteFolder {
-        id: uuid::Uuid::new_v4().to_string(),
-        name: clean,
-        path: rel_path,
-        parent_id: parent_id.map(|s| s.to_string()),
-        locked: false,
-        // A freshly created folder is never sealed, so never session-unlocked.
-        unlocked: false,
-        // Only the reserved Notes root is `is_root`; a user-created folder never is.
-        is_root: false,
-        kind: "note".into(),
-    };
-    state
-        .db
-        .insert_note_folder(&folder, &chrono::Utc::now().to_rfc3339())?;
-    Ok(folder)
-}
-
-/// Rename a note-folder (reuses the meeting-folder rename machinery — folder-id based and
-/// kind-agnostic; it rewrites the `path` + all descendant paths + moves the vault dir). Gated to a
-/// note-folder id so a meeting folder can't be renamed through here.
-#[tauri::command]
-pub fn rename_note_folder(
-    state: State<'_, AppState>,
-    id: String,
-    name: String,
-) -> Result<(), AppError> {
-    if state.db.note_folder_by_id(&id)?.is_none() {
-        return Err(AppError::InvalidArg(format!("no note folder {id}")));
-    }
-    rename_folder_inner(state.inner(), id, name)?;
-    Ok(())
-}
-
-/// Delete a note-folder (reuses the meeting-folder delete machinery). Gated to a note-folder id.
-#[tauri::command]
-pub fn delete_note_folder(state: State<'_, AppState>, id: String) -> Result<(), AppError> {
-    if state.db.note_folder_by_id(&id)?.is_none() {
-        return Err(AppError::InvalidArg(format!("no note folder {id}")));
-    }
-    delete_folder_inner(state.inner(), id)
-}
-
-/// Reparent a note-folder. Gated to a note-folder id; the new parent (if any) must also be a
-/// note-folder. Rewrites the folder's `path` (and every descendant's) so vault export + gating stay
-/// coherent, holding the lifecycle guard so it can't interleave with a lock/unlock.
-#[tauri::command]
-pub fn move_note_folder(
-    state: State<'_, AppState>,
-    id: String,
-    parent_id: Option<String>,
-) -> Result<(), AppError> {
-    move_note_folder_inner(state.inner(), &id, parent_id.as_deref())
-}
-
-/// Inner of [`move_note_folder`] taking `&AppState`.
-pub(crate) fn move_note_folder_inner(
-    state: &AppState,
-    id: &str,
-    parent_id: Option<&str>,
-) -> Result<(), AppError> {
-    let _lifecycle = lifecycle_guard(state);
-    let folder = state
-        .db
-        .note_folder_by_id(id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note folder {id}")))?;
-    // New parent must be a note-folder (or None = a root note-folder under "Notes").
-    let parent_path = match parent_id {
-        Some(pid) => {
-            if pid == id {
-                return Err(AppError::InvalidArg("a folder cannot be its own parent".into()));
-            }
-            state
-                .db
-                .note_folder_by_id(pid)?
-                .ok_or_else(|| AppError::InvalidArg(format!("no parent note folder {pid}")))?
-                .path
-        }
-        None => {
-            state.db.ensure_default_note_folder()?;
-            "Notes".to_string()
-        }
-    };
-    let new_path = format!("{parent_path}/{}", folder.name);
-    if new_path == folder.path {
-        return Ok(()); // no-op reparent.
-    }
-    // A note-folder cannot move under its own descendant (would orphan the subtree). Descendants
-    // have a path prefixed by this folder's path + "/".
-    if parent_path == folder.path || parent_path.starts_with(&format!("{}/", folder.path)) {
-        return Err(AppError::InvalidArg(
-            "cannot move a folder into its own descendant".into(),
-        ));
-    }
-    // Move the vault directory (best-effort) + rewrite this folder's + descendants' paths in the DB.
-    reparent_note_folder_paths(state, id, &folder.path, &new_path, parent_id)?;
-    Ok(())
-}
-
-/// Rewrite `folders.path` for a moved note-folder and EVERY descendant (prefix rewrite), reparent
-/// the row, and move the vault directory on disk (best-effort). Path uniqueness is preserved by the
-/// prefix rewrite (the whole subtree moves as a unit). Kept small + note-scoped (the meeting-folder
-/// rename has its own richer machinery; a note-folder tree is simpler).
-fn reparent_note_folder_paths(
-    state: &AppState,
-    id: &str,
-    old_path: &str,
-    new_path: &str,
-    parent_id: Option<&str>,
-) -> Result<(), AppError> {
-    // Vault dir move first (best-effort; a leftover/absent dir is reconcilable, lost content is not,
-    // but note content lives in the DB — the .md files are re-exportable).
-    if let Some(vault) = vault_path(state) {
-        let vault_root = std::path::Path::new(&vault);
-        let src = assert_in_vault(vault_root, std::path::Path::new(old_path))?;
-        let dst = assert_in_vault(vault_root, std::path::Path::new(new_path))?;
-        if src.exists() {
-            if let Some(parent) = dst.parent() {
-                let _ = std::fs::create_dir_all(parent);
-            }
-            let _ = std::fs::rename(&src, &dst);
-        }
-    }
-    state
-        .db
-        .reparent_note_folder(id, old_path, new_path, parent_id)?;
-    Ok(())
 }
 
 // ── NOTES — selection Brain-assistant (WP4) ──────────────────────────────────────────────────────
